chore: update test utils
